+++ hw06-maven/src/main/java/ru/otus/dto/responses/BooksResponse.java
@@ -2,9 +2,6 @@ package ru.otus.dto.responses;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ru.otus.entities.Author;
-import ru.otus.entities.Book;
-import ru.otus.entities.Genre;
 
 import java.util.List;
 
@@ -15,19 +12,6 @@ public class BooksResponse {
     private final String name;
     private final List<String> authors;
     private final List<String> genres;
-
-    public BooksResponse(Book book) {
-        this.id = book.getId();
-        this.name = book.getName();
-        this.authors = book.getAuthors()
-                .stream()
-                .map(Author::getName)
-                .toList();
-        this.genres = book.getGenres()
-                .stream()
-                .map(Genre::getName)
-                .toList();
-    }
 
     @Override
     public String toString() {
