+++ hw06-maven/src/main/java/ru/otus/entities/Book.java
@@ -33,18 +33,21 @@ public class Book implements Serializable {
             sequenceName = "books_seq")
     @Column(name = "id")
     private Long id;
+
     @Column(name = "name")
     private String name;
+
     @Fetch(FetchMode.SELECT)
     @BatchSize(size = 100)
-    @ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.PERSIST)
+    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.PERSIST, CascadeType.MERGE})
     @JoinTable(name = "books_authors",
             joinColumns = @JoinColumn(name = "books", referencedColumnName = "id"),
             inverseJoinColumns = @JoinColumn(name = "authors", referencedColumnName = "id"))
     private List<Author> authors;
+
     @Fetch(FetchMode.SELECT)
     @BatchSize(size = 100)
-    @ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.PERSIST)
+    @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.PERSIST, CascadeType.MERGE})
     @JoinTable(name = "books_genres",
             joinColumns = @JoinColumn(name = "books", referencedColumnName = "id"),
             inverseJoinColumns = @JoinColumn(name = "genres", referencedColumnName = "id"))
