+++ hw06-maven/src/main/java/ru/otus/repositories/AuthorsRepository.java
@@ -6,6 +6,5 @@ public interface AuthorsRepository extends Repository<Author, Long> {
 
     Author getByName(String name);
 
-    boolean existName(String name);
 }
