+++ hw06-maven/src/main/java/ru/otus/repositories/AuthorsRepositoryImpl.java
@@ -3,13 +3,13 @@ package ru.otus.repositories;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.NoResultException;
 import jakarta.persistence.PersistenceContext;
-import org.springframework.stereotype.Repository;
+import org.springframework.stereotype.Component;
 import ru.otus.entities.Author;
 
 import java.util.List;
 import java.util.Optional;
 
-@Repository
+@Component
 public class AuthorsRepositoryImpl implements AuthorsRepository {
 
     @PersistenceContext
@@ -50,29 +50,12 @@ public class AuthorsRepositoryImpl implements AuthorsRepository {
 
     @Override
     public Author update(Author entity) {
-        entity.getBooks()
-                .stream()
-                .filter(book -> book.getId() == null)
-                .forEach(em::persist);
-
         return em.merge(entity);
     }
 
     @Override
     public void delete(Long id) {
         em.remove(em.find(Author.class, id));
-    }
-
-    @Override
-    public boolean exist(Long id) {
-        try {
-            long count = em.createQuery("select count(a.id) from Author a where a.id = :id", Long.class)
-                    .setParameter("id", id)
-                    .getSingleResult();
-            return count > 0;
-        } catch (NoResultException noResultException) {
-            return false;
-        }
     }
 
     @Override
@@ -86,15 +69,4 @@ public class AuthorsRepositoryImpl implements AuthorsRepository {
         }
     }
 
-    @Override
-    public boolean existName(String name) {
-        try {
-            long count = em.createQuery("select count(a.id) from Author a where a.name = :name", Long.class)
-                    .setParameter("name", name)
-                    .getSingleResult();
-            return count > 0;
-        } catch (NoResultException noResultException) {
-            return false;
-        }
-    }
 }
