+++ hw06-maven/src/main/java/ru/otus/repositories/GenresRepository.java
@@ -5,5 +5,4 @@ import ru.otus.entities.Genre;
 public interface GenresRepository extends Repository<Genre, Long>{
     Genre getByName(String name);
 
-    boolean existName(String name);
 }
