+++ hw06-maven/src/main/java/ru/otus/repositories/Repository.java
@@ -15,7 +15,5 @@ public interface Repository<T, ID> {
 
     void delete(ID id);
 
-    boolean exist(ID id);
-
     int count();
 }
