+++ hw06-maven/src/test/java/ru/otus/repositories/GenresRepositoryImplTest.java
@@ -137,49 +137,6 @@ class GenresRepositoryImplTest {
     }
 
     @Test
-    @DisplayName("при проверке на существование должен выполнить sql запрос к БД и вернуть логическое значение")
-    void existTrue() {
-        when(
-                entityManager.createQuery(
-                        eq("select count(g) from Genre g where g.id = :id"),
-                        eq(Long.class)))
-                .thenReturn(countQuery);
-        when(countQuery.setParameter("id", GENRE_ID)).thenReturn(countQuery);
-        when(countQuery.getSingleResult()).thenReturn(TEST_INT);
-
-        var result = repository.exist(GENRE_ID);
-
-        assertTrue(result);
-        verify(entityManager, times(1))
-                .createQuery(
-                        eq("select count(g) from Genre g where g.id = :id"),
-                        eq(Long.class));
-        verify(countQuery, times(1)).setParameter(eq("id"), eq(GENRE_ID));
-        verify(countQuery, times(1)).getSingleResult();
-    }
-
-    @Test
-    @DisplayName("при проверке на существование должен выполнить sql запрос к БД и вернуть логическое значение")
-    void existFalse() {
-        when(entityManager.createQuery(
-                "select count(g) from Genre g where g.id = :id",
-                Long.class))
-                .thenReturn(countQuery);
-        when(countQuery.setParameter(eq("id"), eq(GENRE_ID))).thenReturn(countQuery);
-        when(countQuery.getSingleResult()).thenReturn(0L);
-
-        var result = repository.exist(GENRE_ID);
-
-        assertFalse(result);
-        verify(entityManager, times(1))
-                .createQuery(
-                        eq("select count(g) from Genre g where g.id = :id"),
-                        eq(Long.class));
-        verify(countQuery, times(1)).setParameter(eq("id"), eq(GENRE_ID));
-        verify(countQuery, times(1)).getSingleResult();
-    }
-
-    @Test
     @DisplayName("при получении количества выполнить sql запрос к БД и вернуть число")
     void count() {
         when(
@@ -200,56 +157,11 @@ class GenresRepositoryImplTest {
     }
 
     @Test
-    @DisplayName("при проверке на то что Автор уже существует должен выполнить sql запрос к БД и вернуть логическое значение")
-    void existNameTrue() {
-        when(
-                entityManager.createQuery(
-                        eq("select count(g) from Genre g where g.name = :name"),
-                        eq(Long.class)))
-                .thenReturn(countQuery);
-        when(countQuery.setParameter("name", GENRE_NAME)).thenReturn(countQuery);
-        when(countQuery.getSingleResult()).thenReturn(TEST_INT);
-
-        var result = repository.existName(GENRE_NAME);
-
-        assertTrue(result);
-        verify(entityManager, times(1))
-                .createQuery(
-                        eq("select count(g) from Genre g where g.name = :name"),
-                        eq(Long.class));
-        verify(countQuery, times(1)).setParameter(eq("name"), eq(GENRE_NAME));
-        verify(countQuery, times(1)).getSingleResult();
-    }
-
-    @Test
-    @DisplayName("при проверке на то что Автор уже существует должен выполнить sql запрос к БД и вернуть логическое значение")
-    void existNameFalse() {
-        when(entityManager.createQuery(
-                "select count(g) from Genre g where g.name = :name",
-                Long.class))
-                .thenReturn(countQuery);
-        when(countQuery.setParameter(eq("name"), eq(GENRE_NAME))).thenReturn(countQuery);
-        when(countQuery.getSingleResult()).thenReturn(0L);
-
-        var result = repository.existName(GENRE_NAME);
-
-        assertFalse(result);
-        verify(entityManager, times(1))
-                .createQuery(
-                        eq("select count(g) from Genre g where g.name = :name"),
-                        eq(Long.class));
-        verify(countQuery, times(1)).setParameter(eq("name"), eq(GENRE_NAME));
-        verify(countQuery, times(1)).getSingleResult();
-    }
-
-    @Test
     @DisplayName("обрабатывает NoResultException")
     void existNameNoResult() {
         when(entityManager.createQuery(any(), any()))
                 .thenThrow(new NoResultException());
 
-        assertFalse(repository.existName(GENRE_NAME));
-        assertFalse(repository.exist(GENRE_ID));
         assertEquals(0, repository.count());
     }
 }
